Fix missing output clear when reading text array
If an ASCII FBX file has both tangents and normals in it, then currently the parsed geometry normals wrongly contain tangent data. This seems to be caused by nothing ever clearing `tmp->v3` array, so while parsing the file first the tangents are written into it, and then the normals appended at the end.